requirements.txt
Pillow
pynput
pystray
pytesseract
